requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
apscheduler==3.10.4
aiosqlite==0.20.0
python-dotenv==1.0.1
